requirements.txt

fastapi>=0.110
uvicorn[standard]>=0.23
httpx>=0.24
python-dotenv>=1.0
